apps/file-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery[redis]==5.4.0
redis==5.0.4
pillow==10.3.0
boto3==1.34.88
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
alembic==1.13.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
ruff==0.4.2
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
httpx==0.27.0
